feat(RemittanceAmount) : RemittanceAmount는 송금액이다. RemittanceAmount와 Exchange를 통하여 수취금액을 구할 수 있도록 한다. RemittanceAmount는 10000을 넘기면 에러를 발생시키도록 한다.

+++ src/main/java/com/youth/exchange/domain/Exchange.java
@@ -15,14 +15,14 @@ public class Exchange {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String pair;
-    private Float rate;
+    private Double rate;
     private LocalDateTime renewalTime;
 
-    public Exchange(String pair, Float rate, LocalDateTime renewalTime) {
+    public Exchange(String pair, Double rate, LocalDateTime renewalTime) {
         this(null, pair, rate, renewalTime);
     }
 
-    public Exchange(Long id, String pair, Float rate, LocalDateTime renewalTime) {
+    public Exchange(Long id, String pair, Double rate, LocalDateTime renewalTime) {
         this.id = id;
         this.pair = pair;
         this.rate = rate;
